requirements.txt
pycryptodome==3.23.0
setuptools==80.9.0
tqdm==4.67.1
wheel==0.45.1
